Add files via upload
This code is about the Rule of the Greatest Number.
You can put the input of three parameters in first line. (N M K)
In the next line, you also enter an array which contains N integers.
The rule of Greatest Number is deriving from the biggest sum by adding some of N integers in M times in your input array.
This rule can add the number in particular index up to K times.   

It's the typical problem of Greedy Algorithm .