Add files via upload
369 Game is an Korean Number Game.
The Rule is as follow.
You and your friends say a number in ascending order from 1 taking turns until the game ends.
When the number contains number 3,6,9 (for example 33, 26, 19..) the person who say that number have to clap his hands.
If the number contains the flag(3,6,9) twice then the person have to clap twice. (for instance, 33,66,..)
This game ends when someone makes a mistake of not clapping or clapping when not clapping.